Implement SyncType
toSyncType itself was removed and was expanded into wrapFsync and
wrapFsyncdir.

The Enum instance was not ported from HFuse because the only use of it
seemed to convert from CInt to SyncType using `toEnum`, which is unsafe
because fuse.h only specifies that "If the datasync parameter is
non-zero..." but `toEnum` assumes it is 1.