Change the build-type to Configure to detect fuse_operations.lseek
lseek was recently added to libfuse at 3.8.0 released as early as
2019-11-03.
Some Linux distributions provide an older version of it, causing
peek/poke operation to fail.

The build-type was converted to Configure and an autoconf script
(configure.ac) was added to detect whether the installed libfuse has
lseek or not.