Extend the lower bound of base to 4.13
To let ghc-8.8 compile this package.

I didn't test ghc-8.6 and older because the Docker image of Debian which
is the base image of haskell:8.6 is too old to install libfuse3-dev.
Perhaps there is a better way to test the other versions of ghc than the
official Docker images?