Use actions/setup-haskell@v1
I misunderstood that "v1" meant "the first release of v1" (i.e. "v0")
but in fact it meant "the latest release of v1", much like the tags used
in Docker.

v1.1.3, the latest release fixes the deprecation warnings around
`add-path` and I don't care which a particular version is used.

(but... in that case, shouldn't "v1.1" also be updated to point to
"v1.1.3" ???)