Allow bytestring-0.11
Since unix-2.7.2.2 now allows bytestring-0.11, libfuse3 is buildable
with it.

Note that certain dependencies of the examples and the benchmark don't
still allow bytestring-0.11.

Confirmed that the library and the tests are buildable and the tests
pass.

Too bad that my cabal-bounds trick couldn't detect this automatically.
Perhaps it's because cabal (? Or ghc??) picks up one bundled with ghc
whether I pass --allow-newer or not.