Change the type of FileStat.blockCount from CBlkSize to CBlkCnt
Since it corresponds to st_blocks, its type should be the one that
corresponds to blkcnt_t.

The URL to the GitHub PR was removed because it is no longer relevant;
unix-2.8.0.0 is out. Also it was misleading because the comment looked
like as if the referenced PR used CBlkSize as the type of st_blocks.
I have no idea about the truth because the line of the code was added
before the git repository was initialized.